splitted files and removed operators
